Feature/nick2

+++ services/transactions/deleteTransaction.js
@@ -1,10 +1,25 @@
-const { NotFoundError } = require('../../helpers');
-const { Transaction } = require('../../models');
+const { NotFoundError, checkUserBalance } = require('../../helpers');
+const { Transaction, User } = require('../../models');
 
-const deleteTransaction = async id => {
-  const result = await Transaction.findByIdAndRemove(id);
-  if (!result) {
-    throw new NotFoundError(`Transaction with id=${id} not found`);
+const deleteTransaction = async ({ id, ownerId }) => {
+  const user = await User.findById(ownerId);
+  const userBalance = user.balance;
+  const userTransaction = await Transaction.findById(id);
+  const userTransactionSum = userTransaction.sum;
+  const userTransactionType = userTransaction.type;
+  const newUserBalance =
+    userTransactionType === 'income'
+      ? user.balance - userTransactionSum
+      : user.balance + userTransactionSum;
+
+  if (checkUserBalance(newUserBalance)) {
+    await User.findByIdAndUpdate(
+      ownerId,
+      { balance: newUserBalance },
+      { new: true },
+    );
+    const result = await Transaction.findByIdAndRemove(id);
+    return result;
   }
 };
